Add footer css and complete website
The footer needed a different background color and
white text. It also needed to be flexed and centered.
This is accomplished.

A small fix with the body content has been implemented.
There is now no padding or margin to the whole of the
body allowing for the content to fill out the top,
bottom, and sides of the page.